% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN.R
\name{filterReads}
\alias{filterReads}
\title{filter out reads based on mapq and LQ20}
\usage{
filterReads(reads, baseQ13 = NULL, mapq = 20)
}
\arguments{
\item{reads}{The sequence of reads, a GAlignments object.
Assumed to have a column qual of PhredQuality objects or a column mapq of mapping quality.}

\item{baseQ13}{Choose reads with the number of at most bad base quality (< 13). default NULL}

\item{mapQ}{Remove all reads with quality less than mapq. Default 20. Set to NULL to keep
all reads.}
}
\value{
The altered reads.
}
\description{
filter out reads based on mapq and LQ20
}
