% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN.R
\name{getReferenceSeq}
\alias{getReferenceSeq}
\title{Extract reference sequence}
\usage{
getReferenceSeq(gr, genome = BSgenome.Hsapiens.UCSC.hg38)
}
\arguments{
\item{gr}{a GRange object indicate the range to retrieve}

\item{genome}{a character string for fasta file location, or a BSgenome reference object}
}
\description{
Get the segments of reference sequence from BSgenome library or a fastq file
}
