% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reAlign.R
\name{reAlign}
\alias{reAlign}
\title{Local re-align reads}
\usage{
reAlign(var, bam, genome = BSgenome.Hsapiens.1000genomes.hs37d5, ...)
}
\arguments{
\item{var}{A GRanges object with REF and ALT in metadata columns.}

\item{bam}{a character vector of BAM files.}

\item{genome}{The path of the reference genome file, or a BSgenome
reference object.}

\item{...}{More options for verIndel.}
}
\description{
To align reads to reference and alternative sequences based on
given indels/variants.
}
