% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN.R
\name{RefAltAln}
\alias{RefAltAln}
\title{Construct alternative reference sequence}
\usage{
RefAltAln(vrange, reads, check.range,
  genome = BSgenome.Hsapiens.UCSC.hg38)
}
\description{
To construct an alternative referecen sequence.
This depend on:
    1. the standard and alternative sequence
    2. check region, in particular the start position
    3. the standard sequence
}
\details{
ISSUE: in repeatRegion()/repReg() we had calculated alta. relationship?
}
