% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verIndel.R
\name{verIndel}
\alias{verIndel}
\title{Verify Indel alignments}
\usage{
verIndel(range, Ref = NULL, Alt = NULL, bam,
  genome = BSgenome.Hsapiens.1000genomes.hs37d5, mapq = 20,
  baseQ13 = NULL, trim.softclip = FALSE)
}
\description{
Verify Indel alignments
}
