% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewAligns.R
\name{viewAligns}
\alias{viewAligns}
\title{viewAligns}
\usage{
viewAligns(grange, bams = NULL, aligns = NULL, genome = "hg19",
  window = 50, bamParam = list(), ...)
}
\arguments{
\item{grange}{A GRanges object.}

\item{bams}{PATHs of the BAM files.}

\item{aligns}{A `GAlignments` or `GAlignmentsList` object.}

\item{genome}{A character string, one of "hg38", "hg19", "mm10",
"tair10".}

\item{window}{The windown to extend for the GRanges.}

\item{bamParam}{The option list for `ScanBamParam`.}

\item{...}{Other options for `igvR`.}
}
\description{
view BAM files in a given range using igvR.
}
