% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicFiles.R
\name{VerINDEL}
\alias{VerINDEL}
\title{VerINDEL}
\usage{
VerINDEL(vranges, bams, outDir,
  genome = BSgenome.Hsapiens.1000genomes.hs37d5, reduceBy = c("file",
  "range"), ...)
}
\description{
verIndel wrapped by GenomicFiles
}
